backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
python-dotenv==1.0.1
pydantic-settings==2.4.0
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
argon2-cffi==23.1.0
python-jose==3.3.0
celery==5.4.0
redis==5.0.8
Pillow==10.4.0
pgvector==0.2.5
torch==2.4.0+cpu
torchvision==0.19.0+cpu
facenet-pytorch==2.5.3
